+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # WIP
 - improve slider UI
+- uncheck "lock" and "compose" when switching project and group
 - fix slider index error
 
 # v1.0.0 (2026-03-31)